Refactor: move d400 xml builder code into own document class
