+++ app/lib/submission_builder/ty2024/states/nc/nc_return_xml.rb
@@ -4,22 +4,6 @@ module SubmissionBuilder
     module States
       module Nc
         class NcReturnXml < SubmissionBuilder::StateReturn
-          FILING_STATUS_OPTIONS = {
-            head_of_household: 'HOH',
-            married_filing_jointly: 'MFJ',
-            married_filing_separately: 'MFS',
-            qualifying_widow: 'QW',
-            single: "Single",
-          }.freeze
-
-          STANDARD_DEDUCTIONS = {
-            head_of_household: 19125,
-            married_filing_jointly: 25500,
-            married_filing_separately: 12750,
-            qualifying_widow: 25500,
-            single: 12750,
-          }.freeze
-
           private
 
           def build_xml_doc_tag
@@ -41,7 +25,7 @@ module SubmissionBuilder
           def supported_documents
             supported_docs = [
               {
-                xml: nil,
+                xml: SubmissionBuilder::Ty2024::States::Nc::Documents::D400,
                 pdf: PdfFiller::NcD400Pdf,
                 include: true
               },
@@ -50,32 +34,6 @@ module SubmissionBuilder
             supported_docs += combined_w2s
 
             supported_docs
-          end
-
-          def documents_wrapper
-            xml_doc = build_xml_doc("FormNCD400") do |xml|
-              xml.ResidencyStatusPrimary true
-              xml.ResidencyStatusSpouse true if @submission.data_source.filing_status_mfj?
-              xml.FilingStatus filing_status
-              if @submission.data_source.filing_status_mfs?
-                xml.MFSSpouseName @submission.data_source.direct_file_data.spouse_name
-                xml.MFSSpouseSSN @submission.data_source.direct_file_data.spouse_ssn
-              end
-              if @submission.data_source.filing_status_qw?
-                xml.QWYearSpouseDied Date.parse(@submission.data_source.direct_file_data.spouse_date_of_death).year
-              end
-              xml.FAGI @submission.data_source.direct_file_data.fed_agi
-              xml.NCStandardDeduction standard_deduction
-            end
-            xml_doc.at('*')
-          end
-
-          def filing_status
-            FILING_STATUS_OPTIONS[@submission.data_source.filing_status]
-          end
-
-          def standard_deduction
-            STANDARD_DEDUCTIONS[@submission.data_source.filing_status]
           end
         end
       end
